Fixed VS compilation (mostly by including winsock2.h before libcurl). Also replaced binary numbers (non-standard) with hex numbers, and a few minor cosmetic changes.
git-svn-id: svn+ssh://svn.code.sf.net/p/supertuxkart/code/main/branches/hilnius@13367 178a84e3-b1eb-0310-8ba1-8eac791a3b58

+++ src/online/servers_manager.cpp
@@ -104,7 +104,7 @@ namespace Online{
             {
                 addServer(new Server(*servers_xml->getNode(i)));
             }
-            m_last_load_time = Time::getRealTime();
+            m_last_load_time = (float)Time::getRealTime();
         }
         m_info_message = input->getInfo();
         //FIXME error message
